Move api middleware module to sub dir
