+++ workers/loc.api/sync/data.inserter/api.middleware/api.middleware.handler.after.js
@@ -6,7 +6,7 @@ const {
   inject
 } = require('inversify')
 
-const TYPES = require('../../di/types')
+const TYPES = require('../../../di/types')
 const {
   addPropsToResIfExist,
   getFlagsFromLedgerDescription
