+++ workers/loc.api/sync/data.inserter/api.middleware/helpers/index.js
@@ -1,0 +1,13 @@
+'use strict'
+
+const addPropsToResIfExist = require(
+  './add-props-to-res-if-exist'
+)
+const getFlagsFromLedgerDescription = require(
+  './get-flags-from-ledger-description'
+)
+
+module.exports = {
+  addPropsToResIfExist,
+  getFlagsFromLedgerDescription
+}
