+++ workers/loc.api/sync/data.inserter/api.middleware/index.js
@@ -9,7 +9,7 @@ const {
   inject
 } = require('inversify')
 
-const TYPES = require('../../di/types')
+const TYPES = require('../../../di/types')
 
 class ApiMiddleware {
   constructor (
